pyproject.toml
[project]
name = "cancer-digital-twins"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aisuite[all]>=0.1.0",
    "clickhouse-driver>=0.2.9",
    "streamlit>=1.35.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
